Remove MSan unpoisoning of memory writes
Prior to supporting MemorySanitizer instrumentation of Reactor routines,
false positives in C++ code were avoided by marking all memory that was
written to by the Reactor routines as unpoisoned.

Now this can be removed to avoid false negatives, i.e. uses of
uninitialized data either in the Reactor routine or originating from it.

Bug: b/155148722
Change-Id: Ia35e3cdc1a60ba44045869884c64e9875c030291