Set LLVM_ENABLE_ABI_BREAKING_CHECKS for Debug builds only
LLVM_ENABLE_ABI_BREAKING_CHECKS was disabled everywhere except for
Fuchsia. It triggered llvm::Error::assertIsChecked() in the new ORCv2
code because we're supposed to check Expected<> didn't return an error
by casting it to a bool, which we only do in Debug builds.

Since this option costs performance in Release builds and we don't want
to abort on errors (especially benign ones), it should only be enable
for Debug builds. This is also the default behavior for standalone
builds of LLVM:
https://llvm.org/docs/ProgrammersManual.html#abi-breaking-checks

Bug: b/171236524
Change-Id: If886fe9ef61d0f5392949d34a63ed3d7de7aa916