Subzero: fix non-deterministic stack layout and code-gen
Using std::unordered_set to hold unmaterialized variables meant that
when materializeAll was called, the order of variables would differ
between runs of the same executable on the same target. Instead, use a
std::vector, which not only guarantees order-consistency across runs,
but also means we submit variables in the order they are declared, which
may be preferable by the JIT backend.

Note that this also affects LLVM; however, LLVM likely already performs
stack variable reordering, so the effects are minimized. Nonetheless,
this change likely also fixes determinism when using LLVM.

Bug: TODO
Change-Id: I5698085ab96663b3c9a2fae4f01b1c1b3aac8b4b