Add baseline VK_EXT_device_memory_report support
This initial implementation will only emit device memory report
callbacks upon vkAllocateMemory and vkFreeMemory.

Memory backing the swapchain images will be attributed to the
VkDeviceMemory object invisible to the application, but it's fine to
leave it as it is.

c++17 is required for including <ui/GraphicBuffer.h>.

VkDeviceMemoryExternalAndroid.hpp shouldn't be included in CMakeLists.

Bug: b/158094132
Change-Id: I519b8c7efed15924e2ac8dbb8ab806d44f4fc1ed