Only lower MSan codegen optimization for LLVM JIT
The large slowdown in codegen for MemorySanitizer instrumented Reactor
code only affects the LLVM backend, and the workaround of not performing
codegen optimizations is also LLVM specific. So only override the
optimization level for LLVM for MSan builds. Previously it was done in
Nucleus, but this abstract API affects other backends too, and this
depended on rr::Optimization::Level::None to correspond with
llvm::CodeGenOpt::None which might not always remain the case.

Bug: b/155148722
Change-Id: I2499935b1ddbd3cefd4b4497094c70ce284b170c