Update Kokoro's clang-format to version 11.0.1
Downloaded from:
https://github.com/llvm/llvm-project/releases/tag/llvmorg-11.0.1

It fixes obeying PointerAlignment for the ellipsis in variadic templates
(https://reviews.llvm.org/D83564), so patch all the files that use them.

clang-format 11 also introduces support for IndentCaseBlocks
(https://reviews.llvm.org/D72276), but leave that unused for now. The
default (false) does not change formatting.

Note that Chromium's buildtools comes with a version of clang-format
which claims to be 11.0.0 but is actually older and does not produce
the same formatting as the official release version:
http://crbug.com/1150875. Use binaries from the download link above
instead.

Bug: b/144825072
Change-Id: I1c9a680f5619926db1ed93afea94d8942926af37