ReactorUnitTests: use test name as Routine names
Rather than using "one" everywhere, we now use the test's name as the
Routine name. This will be used in a future change to generate useful
file names for JIT routine asm listings.

Note that Coroutine does not yet expose a way to pass in a Routine name,
so for now, they are still hard-coded to "coroutine". Will be fixed in a
future change.

Bug: b/174358505
Change-Id: I18e118674b024575cbe4d89100f56f6f45791f6c