Coroutine: allow a name to be passed in for routine generation
Coroutine's call operator accepts variable arguments to be passed to the
underlying entry routine, so we cannot simply add a "const char* name"
parameter. Instead, we wrap it up in a strongly-typed Name class and add
overloads that accept it as a first argument. Note that we may decide to
force a name to always be passed in, as we do for Function, as it's
useful for debugging.

- ReactorUnitTests Coroutine tests now pass in the test name
- ComputeProgram now passes in "ComputeProgram" as a name

Bug: b/174358505
Change-Id: I899e8aab5dfd5d461bc7de0e54bc0e2b72e4b4a9