Coroutine: allow a name to be passed in for routine generation
Add an optional parameter to Coroutine::finalize() to give the underlying
routine a name. To support variadic args, we use a strongly-typed Name class.
Note that giving a name to Coroutines is optional, but we may decide to force
one to be given, as we do for Functions, as it's useful for debugging.

- ReactorUnitTests Coroutine tests now pass in the test name
- ComputeProgram now passes in "ComputeProgram" as a name

Bug: b/174358505
Change-Id: I899e8aab5dfd5d461bc7de0e54bc0e2b72e4b4a9