Fix crash when using rr::Array with ENABLE_RR_DEBUG_INFO
We can no longer call materialize() from any constructor since it is now
a virtual function as of swiftshader-cl/45148. When we do this for
rr::Array, we don't alloca enough space, which results in runtime
crashes due to generated code that writes to invalid stack locations.

I spoke to bclayton@ about why this materialize() call was needed in the
LValue constructor, and he said debug info on values (not allocas) never
worked well for him. I tested ReactorUnitTests.PrintPrimitiveTypes and
PrintReactorTypes without the materialize call, and they pass. Indeed,
the rr::Print code reads from RValue::value(), and if this didn't work
properly, I think we'd have problems with RValues in general. Am
removing this, and if we see weird print issues in the future, we can
revisit.

Bug: b/174160049
Change-Id: I83cec95a74f32ef7276ed60f979a4c32f571eb8a