Revert "Eliminate the array size from rr::Variable"
Fixes crash when using rr::Array with ENABLE_RR_DEBUG_INFO. This is
because when the LValue ctor calls materialize, this would end up
calling the virtual allocate() function, which would not call the
overridden version in rr::Array as it was not fully constructed yet.
Reverting the original change that added the virtual allocate().

This reverts commit 5e16bc45a1f7540f697d8f651e023025ee83cdc2.

Bug: b/174160049
Change-Id: I83cec95a74f32ef7276ed60f979a4c32f571eb8a