Add decorate operations new to SPIR-V 1.4
OpDecorateId was added to SPIR-V 1.2, but was missing from SwiftShader.
OpDecorateString and OpMemberDecorateString were added in SPIR-V 1.4 as
a part of the SPV_GOOGLE_decorate_string extension being integrated as
a core part of SPIR-V 1.4.

All of the decorations for these operations are to support HLSL at a
level before the graphics driver is involved, to include debug
information in the SPIR-V code, or to inform the SPIR-V compiler about
potential optimizations that SwiftShader currently doesn't implement.

Test: dEQP-VK.spirv_assembly.instruction.spirv1p4.hlsl_functionality1.*
Test: dEQP-VK.spirv_assembly.instruction.spirv1p4.uniformid.*
Bug: b/174693410
Change-Id: I7a22b335da9f5566e3d08869a80ea6701f035155