Disable SPIR-V validation during pipeline compilation
Validation is costly and it's the applications' responsibility to send
valid SPIR-V to the ICD, so we shouldn't do it in Release builds.

Note that while in Debug builds we already perform validation during
creation of the shader module (to catch issues even sooner), we should
validate again during pipeline compilation to ensure we (or the tooling
we use) didn't alter the SPIR-V to become invalid.

Bug: b/158228522
Change-Id: I5b8153f0397c609d139e37909868b5836d57d40e