Implement VK_KHR_timeline_semaphore
Timeline semaphores are like regular binary semaphores, except instead
of tracking a single boolean payload, they track a monotonically
increasing uint64_t. Threads and the host can wait for that payload to
become a certain value. All threads waiting for the same value should
unblock together.

The device and the host increase the payload by signalling it with a new
value.

This is a naive implementation of timeline semaphores that likely
suffers from race conditions and wasted CPU cycles, but passes dEQP
tests. Each waiting thread when it gets signaled checks if the payload
matches its expected value, then signals another thread to do the same.

A better implementation would only allow the threads waiting on the
current payload value to receive the signal.

Change-Id: Iafd95b8eac11beea21c00df87ca7ca1e51c77c64
Bug: b/147738281
Test: dEQP-VK.*