Implement non-void indirect Reactor Call() support
Previously the return value of the function called through the
Pointer<Byte> argument was not returned from the rr::Call(). This
change implements proper support for both non-void and void returning
function signatures.

This change also adds a unit test which demonstrates the use of a
'trampoline', which makes use of the new non-void indirect call.

Bug: b/175830790
Change-Id: I100588c6319d0c45797c48a068e1e46fa9f441ba