Make texture sampling functions more reusable
TL;DR: Reduced ImageInstruction from 15 bits to 9 bits

This CL reduces the number (and size) of image instruction
parameters used as part of the cache key. In order to do
so, input locations have been made constant, removing the
need for specific parameters to figure out where each input
is located within the input array.

These parameters were removed:
- coordinates (3 bits)
- grad (2 bits)

This parameter was reduced in size:
- offset (2 bits -> 1 bit, we only care about whether or
          not there's an offset)

The downside of this is that the input array is always read
in it's entirety by the texture sampling function, which
seems like a reasonable trade-off for increasing our
ability to pre-compile or even inline them.

Bug: b/168118740
Change-Id: I2904b55cfbf295b1cd46735b610e013dd1b4059c