Log the use of Vulkan result codes
It is useful to known the return values of Vulkan entry functions when
tracing, and to have a central place to put a breakpoint.

Note that this will report any usage of the VkResult enums, including
for example assigning VK_SUCCESS to a local variable before updating it
with another result code. The last logged code is not necessarily the
one returned from the function.

Bug: b/139528621
Change-Id: I8dfc915b535f9eeb594a107c89fedd49863b3e9a