Make MemoryMapper a JITRoutine member
Previously a single MemoryManager with static lifetime was shared
between all JITRoutine instances. This causes a destruction order
issue when it is destroyed at shutdown before atexit() destroys all
routines.

Note that the objectLayer member holds a reference to the memoryMapper
member, and calls it during its destruction, so the latter must be
declared first to be destroyed last.

Bug: b/216348016
Change-Id: Icd85e2687c652f45d78b9d48ed481c020e9cb6f3