Define static constexpr members in .cpp files
Members of classes that are declared as static constexpr are not defined
as part of their declaration. Thus they must be defined elsewhere. And
they can't be defined in the header file, or there will be multiple
definitions for each file that includes the header.

So define these contexprs in Constants.cpp

Bug: b/177220369
Change-Id: I9ffc63867948bc72061ddb8b0b586a46d840196c