Move required interpolation logic to SpirvShader
Some of the interpolation logic is required to properly implement
the GLSLstd450Interpolate* operations. In order to prevent code
duplication and to make sure the logic is always identical between
the regular fragment input interpolation and the one performed by
the GLSLstd450Interpolate* operations, the interpolateCentroid()
function was moved to SpirvShader and renamed with a more generic
name, interpolateXY(), as it will be used to interpolate all modes
(centroid, sample and offset).

Bug: b/171415086
Change-Id: Ifbffa2b395a0b8c22fa120b36643db05f867a306