Implement GLSLstd450Interpolate* functions
This cl adds an implementation for:
- GLSLstd450InterpolateAtCentroid
- GLSLstd450InterpolateAtSample
- GLSLstd450InterpolateAtOffset

These functions essentially replicate the behavior of
regular interpolants in the fragment shader processing.

A specific extra difficulty encountered here is detecting
which kind of pointer offset we are dealing with. Pointer
offsets might be caused by [] operators being used on a
vector or on an array (possibly an array of vectors). This
distinction is important as it impacts what interpolant
offsets point to. Note that there's missing coverage in
dEQP-VK for interpolant arrays and this was caught with
SwANGLE tests (a dEQP-VK issue has been logged, See:
https://gitlab.khronos.org/Tracker/vk-gl-cts/-/issues/2733).

Bug: b/171415086
Change-Id: Id7c4c931918ba172d00da84655051445b110d3a9