Flip samples back to their correct position
SwANGLE tests detected that SwiftShader's sample positions were 180
degree flipped from Vulkan's exact sample locations. Most multisampling
tests we've run with SwiftShader so far relied on resolved outputs,
so having the samples in the wrong order didn't matter, as long as we
had the correct 4 samples (which we do, since the samples are
essentially rotated by 90 degrees around the center of the pixel from
each other).

Bug: b/170962186
Bug: b/141380498
Change-Id: I65552863e1fe38b321b84995e038984488512e1c