Finalize Reactor Function creation at Routine acquisition
This enables generating multiple functions within the same scope, as
excercised by the Trampoline unit test.

Also adds a unit test for functions which are never finalized (useful
for when a code path is not supported and we want to fall back to a C++
implementation).

Also adds a unit test for the common pattern of deriving from Function<>
and using Reactor variables as members.

Bug: b/177999324
Change-Id: I269946e9c8a791eab84ff756c9c588ca8ce26337