Silence dead code warning
Clang produces:
"error: code will never be executed [-Werror,-Wunreachable-code]"
For code following an unconditional `break`. But it hints that:
"note: silence by adding parentheses to mark code as explicitly dead"

Hence the dead code can follow an "if((true)) break;" without producing
the warning/error.

Note that in this instance we could have just removed the dead code, but
I wanted to make it clear that this is an example of an unfinished
Reactor function.

Bug: b/178929463