Force-inline all SPIR-V function calls
SwiftShader does not support function calls in SPIR-V shader code at the
moment, which isn't a big issue due to running the InlineExhaustivePass
optimization pass. This is part of the default set of optimizations
SPIRV-Tools applies. However, it has started to respect the DontInline
flag, and leave calls to functions that declare this flag not inlined.

This change adds a force-inline optimizer option to SPIRV-Tools, and
enables it for SwiftShader.

Note this does not violate the spec since the flag is a "Strong request,
to the extent possible, to not inline the function." but not a demand.
Some GPUs are assumed also not to support function calls and must force-
inline every call too.

Bug: b/172212963
Tests: dEQP-VK.spirv_assembly.instruction.compute.function_control.*
Change-Id: Ia37ca99e0a84b85848bd0d3cbc5a753cc05a80cf
Reviewed-on: https://swiftshader-review.googlesource.com/c/SwiftShader/+/62228
Tested-by: Nicolas Capens <nicolascapens@google.com>
Reviewed-by: Alexis Hétu <sugoi@google.com>