Ignore the DontInline SPIR-V flag
SwiftShader does not support function calls in SPIR-V shader code at the
moment, which isn't a big issue due to running the InlineExhaustivePass
optimization pass. This is part of the default set of optimizations
SPIRV-Tools applies. However, it has started to respect the DontInline
flag, and leave calls to functions that declare this flag not inlined.

This change undoes that by ignoring the flag and force-inlining every
function call.

Note this does not violate the spec since the flag is a "Strong request,
to the extent possible, to not inline the function." but not a demand.
Some GPUs are assumed also not to support function calls and must force-
inline every call too.

Bug: b/172212963
Tests: dEQP-VK.spirv_assembly.instruction.compute.function_control.*
Change-Id: Ic55f0497a95663a374f6ed49b3adaa017845c800