Propagate stores of stack addresses into other local variables
Pointers to the base address of arrays often get stored and then loaded
multiple time. This change adds an optimization pass to replace these
loads with the result of the alloca of the array, and eliminates the
allocas of the pointer variables.

Also add a couple of unit tests for corner cases.

Bug: b/179279298
Change-Id: I784899319bf5360f47c6fbc22fb82134d135dbfc