Eliminate Subzero profiling support
We're never used this functionality, and shouldn't have a need for it.
Profiling information can be collected at the Reactor level or using
a profiler like VTune.

This functionality was the only thing using the `Target` parameter of
`InstIntrinsicCall`, which get in the way for aligning the parameters of
load- and store-like intrinsics with regular `InstLoad` and `InstStore`.

Bug: b/179497998
Change-Id: I5a0ad5ee8e0101f0879a97a1ea01e3efc5bebbe4