Regres: Remove GLES tests from CI test runs
The legacy OpenGL ES implementation is deprecated in favor of SwANGLE,
and we haven't touched the GLES code in many months. These CI tests
consume valuable time, and tend to be flaky due to exhausting the
number of X handles. We still have our 'daily' test runs to provide
detection of regressions, so they can be safely removed from the CI
test runs.

Note that this change can also be (temporarily) reverted as part of a
new change which could use CI testing of dEQP-GLES.

Bug: b/153322216
Change-Id: I52cd2b89c04c95de486d85118edef6460ba82925